requirements.txt
torch>=2.0.0
openai>=1.0.0
python-dotenv>=0.19.0
pydantic>=2.0.0
chromadb>=0.4.0
mcp>=1.0.0
asyncio-pipe>=0.3.0
